Deleted files, swapped bearing to int

+++ CamReg_correction/climb2.h
@@ -31,9 +31,9 @@
 #define COEFF_PROX					0.7		//Determined empirically
 
 
-float imu_bearing(int16_t acc_x, int16_t acc_y);
+int8_t imu_bearing(int16_t acc_x, int16_t acc_y);
 
-float prox_bearing(int prox_front_left, int prox_front_right, int prox_diag_left, int prox_diag_right);
+int8_t prox_bearing(int prox_front_left, int prox_front_right, int prox_diag_left, int prox_diag_right);
 
 void move(float bearing);
